requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0

# NLP & ML
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# LLM Integration
anthropic>=0.25.0
openai>=1.0.0  # Optional: GPT API
requests>=2.31.0  # For Ollama API

# Data Processing
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
jinja2>=3.1.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Caching
redis>=4.5.0

# Optional: API
fastapi>=0.100.0
uvicorn>=0.23.0
